refactor: split vsock-proto frame infrastructure
